chore: seperate types
